Changed vote count in Opinion class to long as Mongo has no concept of uint

+++ BiscuitService.Domain/Models/Opinion.cs
@@ -3,6 +3,6 @@
     public class Opinion
     {
         public string Name { get; set; } = string.Empty;
-        public uint Votes { get; set; }
+        public long Votes { get; set; }
     }
 }
